feat: protect private routes

+++ src/domain/use-cases/client/delete.ts
@@ -1,4 +1,8 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  Inject,
+  Injectable,
+  InternalServerErrorException,
+} from '@nestjs/common';
 import { IClientRepository } from 'src/domain/interfaces/client.repository';
 
 @Injectable()
@@ -9,7 +13,12 @@ export class DeleteClient {
   ) {}
 
   async execute(id: string) {
-    await this.clientRepository.findById(id);
+    const client = await this.clientRepository.findById(id);
+    if (client.deletedAt) {
+      throw new InternalServerErrorException(
+        'This client already has been deleted.',
+      );
+    }
     await this.clientRepository.delete(id);
   }
 }
